+++ src/main/java/com/example/styleflow/entity/Product.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 @Entity // maps this class to a DB table (one row per instance) — NOT a Spring bean
 public class Product {
 
@@ -17,7 +19,7 @@ public class Product {
     private String description;
 
     @Positive // must be > 0
-    private Double price;
+    private BigDecimal price;
 
     @Positive
     private Integer stockQuantity;
@@ -28,7 +30,7 @@ public class Product {
 
     public Product() {} // no-arg constructor required by JPA/Hibernate to build entities
 
-    public Product(Integer id, String name, String description, Double price, Integer stockQuantity, Category category) {
+    public Product(Integer id, String name, String description, BigDecimal price, Integer stockQuantity, Category category) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -44,8 +46,8 @@ public class Product {
     public void setName(String name) { this.name = name; }
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
-    public Double getPrice() { return price; }
-    public void setPrice(Double price) { this.price = price; }
+    public BigDecimal getPrice() { return price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
     public Integer getStockQuantity() { return stockQuantity; }
     public void setStockQuantity(Integer stockQuantity) { this.stockQuantity = stockQuantity; }
     public Category getCategory() { return category; }
